+++ compile/primitive.go
@@ -27,67 +27,117 @@ import (
 	"go.uber.org/thriftrw/wire"
 )
 
-// TypeSpecs for primitive Thrift types.
-var (
-	BoolSpec   TypeSpec = primitiveTypeSpec{Name: "bool", Code: wire.TBool}
-	I8Spec     TypeSpec = primitiveTypeSpec{Name: "byte", Code: wire.TI8}
-	I16Spec    TypeSpec = primitiveTypeSpec{Name: "i16", Code: wire.TI16}
-	I32Spec    TypeSpec = primitiveTypeSpec{Name: "i32", Code: wire.TI32}
-	I64Spec    TypeSpec = primitiveTypeSpec{Name: "i64", Code: wire.TI64}
-	DoubleSpec TypeSpec = primitiveTypeSpec{Name: "double", Code: wire.TDouble}
-	StringSpec TypeSpec = primitiveTypeSpec{Name: "string", Code: wire.TBinary}
-	BinarySpec TypeSpec = primitiveTypeSpec{Name: "binary", Code: wire.TBinary}
+type (
+	BoolSpec struct {
+		nativeThriftType
+
+		Annotations Annotations
+	}
+
+	I8Spec struct {
+		nativeThriftType
+
+		Annotations Annotations
+	}
+
+	I16Spec struct {
+		nativeThriftType
+
+		Annotations Annotations
+	}
+
+	I32Spec struct {
+		nativeThriftType
+
+		Annotations Annotations
+	}
+
+	I64Spec struct {
+		nativeThriftType
+
+		Annotations Annotations
+	}
+
+	DoubleSpec struct {
+		nativeThriftType
+
+		Annotations Annotations
+	}
+
+	StringSpec struct {
+		nativeThriftType
+
+		Annotations Annotations
+	}
+
+	BinarySpec struct {
+		nativeThriftType
+
+		Annotations Annotations
+	}
 )
 
-type primitiveTypeSpec struct {
-	nativeThriftType
+func (*BoolSpec) TypeCode() wire.Type   { return wire.TBool }
+func (*I8Spec) TypeCode() wire.Type     { return wire.TI8 }
+func (*I16Spec) TypeCode() wire.Type    { return wire.TI16 }
+func (*I32Spec) TypeCode() wire.Type    { return wire.TI32 }
+func (*I64Spec) TypeCode() wire.Type    { return wire.TI64 }
+func (*DoubleSpec) TypeCode() wire.Type { return wire.TDouble }
+func (*StringSpec) TypeCode() wire.Type { return wire.TBinary }
+func (*BinarySpec) TypeCode() wire.Type { return wire.TBinary }
 
-	Name string
-	Code wire.Type
-	// TODO(abg): We'll want to expose type annotations here
-}
+func (*BoolSpec) ThriftName() string   { return "bool" }
+func (*I8Spec) ThriftName() string     { return "byte" }
+func (*I16Spec) ThriftName() string    { return "i16" }
+func (*I32Spec) ThriftName() string    { return "i32" }
+func (*I64Spec) ThriftName() string    { return "i64" }
+func (*DoubleSpec) ThriftName() string { return "double" }
+func (*StringSpec) ThriftName() string { return "string" }
+func (*BinarySpec) ThriftName() string { return "binary" }
 
-// TypeCode of the primitive type.
-func (t primitiveTypeSpec) TypeCode() wire.Type {
-	return t.Code
-}
+func (t *BoolSpec) Link(Scope) (TypeSpec, error)   { return t, nil }
+func (t *I8Spec) Link(Scope) (TypeSpec, error)     { return t, nil }
+func (t *I16Spec) Link(Scope) (TypeSpec, error)    { return t, nil }
+func (t *I32Spec) Link(Scope) (TypeSpec, error)    { return t, nil }
+func (t *I64Spec) Link(Scope) (TypeSpec, error)    { return t, nil }
+func (t *DoubleSpec) Link(Scope) (TypeSpec, error) { return t, nil }
+func (t *StringSpec) Link(Scope) (TypeSpec, error) { return t, nil }
+func (t *BinarySpec) Link(Scope) (TypeSpec, error) { return t, nil }
 
-// ThriftName of the primitive type.
-func (t primitiveTypeSpec) ThriftName() string {
-	return t.Name
-}
-
-// ForEachTypeReference of the primitive type.
-func (t primitiveTypeSpec) ForEachTypeReference(func(TypeSpec) error) error {
-	return nil
-}
-
-// Link for primitive types is a no-op because primitive types don't make any
-// references.
-func (t primitiveTypeSpec) Link(Scope) (TypeSpec, error) {
-	return t, nil
-}
+func (*BoolSpec) ForEachTypeReference(func(TypeSpec) error) error   { return nil }
+func (*I8Spec) ForEachTypeReference(func(TypeSpec) error) error     { return nil }
+func (*I16Spec) ForEachTypeReference(func(TypeSpec) error) error    { return nil }
+func (*I32Spec) ForEachTypeReference(func(TypeSpec) error) error    { return nil }
+func (*I64Spec) ForEachTypeReference(func(TypeSpec) error) error    { return nil }
+func (*DoubleSpec) ForEachTypeReference(func(TypeSpec) error) error { return nil }
+func (*StringSpec) ForEachTypeReference(func(TypeSpec) error) error { return nil }
+func (*BinarySpec) ForEachTypeReference(func(TypeSpec) error) error { return nil }
 
 // compileBaseType compiles a base type reference in the AST to a primitive
 // TypeSpec.
-func compileBaseType(t ast.BaseType) TypeSpec {
+func compileBaseType(t ast.BaseType) (TypeSpec, error) {
+	annots, err := compileAnnotations(t.Annotations)
+	if err != nil {
+		return nil, err
+	}
+
 	switch t.ID {
 	case ast.BoolTypeID:
-		return BoolSpec
+		return &BoolSpec{Annotations: annots}, nil
 	case ast.I8TypeID:
-		return I8Spec
+		return &I8Spec{Annotations: annots}, nil
 	case ast.I16TypeID:
-		return I16Spec
+		return &I16Spec{Annotations: annots}, nil
 	case ast.I32TypeID:
-		return I32Spec
+		return &I32Spec{Annotations: annots}, nil
 	case ast.I64TypeID:
-		return I64Spec
+		return &I64Spec{Annotations: annots}, nil
 	case ast.DoubleTypeID:
-		return DoubleSpec
+		return &DoubleSpec{Annotations: annots}, nil
 	case ast.StringTypeID:
-		return StringSpec
+		return &StringSpec{Annotations: annots}, nil
 	case ast.BinaryTypeID:
-		return BinarySpec
+		return &BinarySpec{Annotations: annots}, nil
 	default:
 		panic(fmt.Sprintf("unknown base type %v", t))
 	}